task_4/flask_app/requirements.txt
flask
requests
azure.identity
azure.keyvault.secrets